borrador para inspirarme, argc argv

+++ 01-tabla-nudos-a-km/conversion.c
@@ -1,5 +1,5 @@
 #include "conversion.h"
 
-double conversion(double nudos){
+float conversion(float nudos){
     return nudos*1.852;
 }
