+++ 01-tabla-nudos-a-km/conversion.h
@@ -1,6 +1,6 @@
 #ifdef CONVERSION_H_INCLUDED
 #define CONVERSION_H_INCLUDED
 
-double conversion(double nudos);
+float conversion(float nudos);
 
 #endif
